+++ apps/backend-functions/src/contract.ts
@@ -1,4 +1,4 @@
-import { functions, db, DocumentReference } from './internals/firebase';
+import { functions, db } from './internals/firebase';
 import { ContractDocument, PublicContractDocument, OrganizationDocument } from './data/types';
 import { ValidContractStatuses, ContractVersionDocument } from '@blockframes/contract/contract/+state/contract.firestore';
 import { getOrganizationsOfContract, getDocument, createPublicOrganizationDocument } from './data/internals';
@@ -7,6 +7,7 @@ import { centralOrgID } from './environments/environment';
 import { isEqual, uniqBy, flatten } from 'lodash';
 import { firestore } from 'firebase-admin';
 import { Change } from 'firebase-functions';
+import { DocumentReference } from '@blockframes/firebase-utils';
 
 async function getCurrentVersionId(tx: FirebaseFirestore.Transaction, contractId: string): Promise<number> {
   return (await _getVersionCount(contractId, tx));
