+++ libs/firebase-utils/src/index.ts
@@ -1,4 +1,6 @@
 export * from './lib/firestoreMigrations'
+export * from './lib/firestore'
+export * from './lib/queue'
 export * from './lib/firebase-utils';
 export * from './lib/maintenance';
 export * from './lib/watermark';
