+++ libs/firebase-utils/src/lib/firestore/export.ts
@@ -1,0 +1,1 @@
+export default 5
