+++ libs/firebase-utils/src/lib/types.ts
@@ -8,3 +8,4 @@ export type QueryDocumentSnapshot = admin.firestore.QueryDocumentSnapshot;
 export type QuerySnapshot = admin.firestore.QuerySnapshot;
 export type Transaction = admin.firestore.Transaction;
 export type UserRecord = admin.auth.UserRecord;
+export type CollectionReference = admin.firestore.CollectionReference;
